Script js for Uniting Schemas
Check if you are working on schema.prisma on your branch.
If you are, then no pulling schema from origin/master is needed. Later on, you need to solve conflicts yourself when push to master.
If you are not, then pull schema from origin/master and npm run update-db locally